requirements.txt
pandas~=2.3.0
numpy~=2.0.0
scikit-learn~=1.6.0
xgboost~=2.1.0
catboost~=1.2.0
shap~=0.49.0
matplotlib~=3.9.0
PyYAML~=6.0
joblib~=1.5.0
pytest~=8.4.0
pytest-cov~=5.0.0
